make-package: Allow to specify the source to build the package for
It sometimes happen that the build is broken on unstable due to a
random amount of upstream dependencies failing to build. This makes
our package build impossible to run completely (especially for our
current infra target, stable).

This allows for specifying the target we want the package to be built
for (without having to amend the script to avoid the problem).

The default is to build/upload for all distributions.