Add lister workers services
This includes the npm worker which is not yet published.

The celery workers run under a dedicated swh user otherwise celery
complains quite loudly.

This has a pretty annoying drawback: we cannot pip install swh packages
from volume mounted source in 'develop mode' because vcversioner will
then attempt to write 2 files in the source directory (version.txt and
the _version.py module) in which we do not have write access when
the entrypoint is run from the docker as the swh user.

There is no easy solution for this problem, so for now, we install the
package from sources, bit NOT in develop mode (non editable).

Related to T1443.