docker-compose: Fix missing `start-` prefix on scheduler cli
Prior to this commit, on fresh container update, the listener and runner
scheduler services would not start.

```
swh-scheduler-runner_1        | Waiting for postgresql to start
swh-scheduler-runner_1        | wait-for-it: waiting for swh-scheduler-db:5432 without a timeout
swh-scheduler-runner_1        | wait-for-it: swh-scheduler-db:5432 is available after 0 seconds
swh-scheduler-runner_1        | Starting the swh-scheduler runner
swh-scheduler-runner_1        | wait-for-it: waiting for amqp:5672 without a timeout
swh-scheduler-runner_1        | wait-for-it: amqp:5672 is available after 4 seconds
swh-scheduler-runner_1        | Usage: swh scheduler [OPTIONS] COMMAND [ARGS]...
swh-scheduler-runner_1        | Try "swh scheduler -h" for help.
swh-scheduler-runner_1        |
swh-scheduler-runner_1        | Error: No such command "runner".
swh-docker-dev_swh-scheduler-runner_1 exited with code 2
```

Because we removed the backward compatibility code and forgot to update the
docker env.

Related D1995