vault: Fix tests using the new key names
Names got changed in the vault api backend. Unfortunately, this passed
undetected both by the vault and the webapp tests.

This commit fixes the situation by adapting to from the old names to the new
ones (this should no longer happen and new tests on the vault should make sure
this is caught early).