Refresh save code now statuses periodically
Without the following, as currently, the fresh new build from swh.web 0.0.302 onward,
new save code now requests are not updated anymore through the ui. The following makes
sure the status are updated periodically (15s).

This simply loops through indefinitely and periodically updates in db the save code now
request statuses. As installing a cron was deemed unnecessary [1]

This is the equivalent of what's been deployed in production [2]

[1] D5615

[2] D5616

Related to T3280