docker/conf/swh-web: Ensure to use a new empty database in dev mode
When overriding swh-web sources, default development database will be
used but it corresponds to the local swh-web development database on
the host sytem using sqlite3 backend.

So ensure to use a newly created sqlite3 database located in the
swh-web service container instead to not persist data across docker
environment sessions.