bib/install: disable pip self ugprade
This is both a temporary fix for https://github.com/pypa/pip/issues/10573 ,
which currently breaks entirely our developer seutp, and also a saner approach
to future pip breakages.

As part of this, refactor a bit pip invocation in bin/install to have flags and
invocation logic in a single place.