bin/install: Force legacy behavior of editable installs with recent pip
As mypy does not support editable installs for projects that have a
pyproject.toml and use a build backend that supports PEP 660, force
the legacy behavior of editable installs by passing the configuration
settings "editable_mode=compat" to the build backend.

It enables to remove the pip version restriction in the swh venv and
avoid bad surprises when the mypy pre-commit hook is called when
committing in a swh repository.

See https://github.com/python/mypy/issues/13392 for more details.