docker: Rework the db configuration to add support for multiple dbs
Make it possible to declare several db access on a service (aka have
several entries in pg_service.conf).

Replace usage of most env/xxx-db.env files by direct definitions of env
vars in compose files: this indeed imply a bit of
duplication/synchronization between components of a given service, but
in the end it makes the overall configuration of services easier to
grasp, and most of the time, it consists in a single POSTGRES_DB
declaration.