docker: add dependency on prometheus-statsd-exporter for most services
When only a subset of all the services is deployed, if the statsd
endpoint is not started, all services using statsd will hand for 30s on
each query trying to resolve the name of the statsd service.

So it's a good habit to make sure it's actually started every time.