docker: try to speed up a bit the loading of origins in tests
Spawn a few more swh-loader services if needed, and do not serialize the
creation of loading tasks in the scheduler; i.e. do not wait for a
loading task to be completed before creating the next one.
With several loader workers, we may thus have several origin loading
tasks executed in parallel.