docker: add a 'kafka-create-topics' init service
This is a simple short-lived service dedicated to creating all known
topics. It's using a pyhon script because it seems it's the fastest way
of creating a bunch of topics (other methods, be it using
kafka-topics.sh or using the kafka rest proxy, can only create one topic
at a time, with an execution time of 0.6s to 1.9s per topic...)

Remove all the `create_topics=True` config entries, and provide a helper
tool `wait-for-topic` to make it easier to write entrypoint scripts.

Note that topics are created with default topic config (16 parititons, 1
copy, compact log cleanup policy thanks to the configuration of the
kafka cluster).

Also note that using journal clients with `create_topic` set does not
plays nicely when there are with anonymized/privileged topics, which is
a reason for all this.