docker/tests: Add support for multiple origin urls in the origins fixture
Make it possible to declare one origin to be loaded with multiple urls
and choose the first one to be accessible as url for the test. This can
prevent the tests from failing due to a transient downtime of an origin.

Use this to make the mercurial origin used in test_mirror iand test_graphql
more robust (sr.ht being currently down).